+++ app/src/main/kotlin/com/getstrm/pace/config/AppConfiguration.kt
@@ -8,6 +8,7 @@ data class AppConfiguration(
     val exposeApplicationExceptions: Boolean = false,
     val defaultViewSuffix: String = "",
     val catalogs: List<CatalogConfiguration> = emptyList(),
+    val dataPolicies: DataPoliciesConfiguration = DataPoliciesConfiguration(),
     val processingPlatforms: ProcessingPlatformsConfiguration = ProcessingPlatformsConfiguration(),
 )
 
@@ -19,4 +20,8 @@ data class CatalogConfiguration(
     val userName: String?,
     val password: String?,
     val fetchSize: Int? = 1,
+)
+
+data class DataPoliciesConfiguration(
+    val autoIncrementVersion: Boolean = true,
 )