+++ app/src/main/kotlin/com/getstrm/pace/config/ProcessingPlatformsConfiguration.kt
@@ -31,6 +31,7 @@ data class BigQueryConfiguration(
     val projectId: String,
     val userGroupsTable: String,
     val serviceAccountJsonKey: String,
+    val useIamCheckExtension: Boolean = false,
 ) : ProcessingPlatformConfiguration(id, ProcessingPlatform.PlatformType.BIGQUERY)
 
 data class PostgresConfiguration(