+++ app/src/main/kotlin/com/getstrm/pace/processing_platforms/bigquery/BigQueryClient.kt
@@ -104,7 +104,8 @@ class BigQueryClient(
         getDatabase(databaseId).getChild(schemaId).getChild(tableId)
 
     override suspend fun applyPolicy(dataPolicy: DataPolicy) {
-        val viewGenerator = BigQueryViewGenerator(dataPolicy, config.userGroupsTable)
+        val viewGenerator =
+            BigQueryViewGenerator(dataPolicy, config.userGroupsTable, config.useIamCheckExtension)
         val query = viewGenerator.toDynamicViewSQL().sql
         val queryConfig = QueryJobConfiguration.newBuilder(query).setUseLegacySql(false).build()
         try {